requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
